+++ client/Packages/com.beamable.server/SharedRuntime/Api/Payments/IMicroservicePaymentsApi.cs
@@ -1,9 +1,9 @@
-﻿using Beamable.Common.Api.Payments;
+using Beamable.Common.Api.Payments;
 
 namespace Beamable.Server.Api.Payments
 {
 	public interface IMicroservicePaymentsApi : IPaymentsApi
 	{
-		
+
 	}
 }